Asset generation on build added
