+++ package.json
@@ -47,7 +47,7 @@
   },
   "scripts": {
     "start": "wp-scripts start",
-    "build": "wp-scripts build",
+    "build": "wp-scripts build && node scripts/generate-assets-files.js",
     "format:js": "wp-scripts format-js",
     "lint:js": "wp-scripts lint-js",
     "lint:js:fix": "wp-scripts lint-js --fix",
