+++ scripts/paths.js
@@ -14,8 +14,9 @@ const resolvePlugin = relativePath => path.resolve( pluginDir, relativePath )
 // Config after eject: we're in ./config/
 module.exports = {
 	dotenv: resolvePlugin( ".env" ),
+	pluginDir,
 	pluginSrc: resolvePlugin( "src" ), // Plugin src folder path.
 	pluginBlocksJs: resolvePlugin( "src/blocks.js" ),
 	yarnLockFile: resolvePlugin( "yarn.lock" ),
-	pluginDist: resolvePlugin( "." ), // We are in ./dist folder already so the path '.' resolves to ./dist/.
+	pluginDist: resolvePlugin( "dist" ), // We are in ./dist folder already so the path '.' resolves to ./dist/.
 }
